requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose
pydantic
